+++ lib/common/common_methods.dart
@@ -1,9 +1,9 @@
 import 'package:provider/provider.dart';
-import 'package:retroshare/model/chat.dart';
-import 'package:retroshare/model/identity.dart';
+
 import 'package:retroshare/provider/friends_identity.dart';
 import 'package:retroshare/provider/Idenity.dart';
 import 'package:retroshare/provider/room.dart';
+import 'package:retroshare_api_wrapper/retroshare.dart';
 
 int getUnreadCount(context, Identity identity) {
   return Provider.of<RoomChatLobby>(context, listen: false).distanceChat != null
