--- lib/model/account.dart
@@ -1,8 +1,0 @@
-class Account {
-  String locationId;
-  String pgpId;
-  String locationName;
-  String pgpName;
-
-  Account(this.locationId, this.pgpId, this.locationName, this.pgpName);
-}
