--- lib/model/location.dart
@@ -1,15 +1,0 @@
-class Location {
-  String rsPeerId;
-  String rsGpgId;
-  String accountName;
-  String locationName;
-  bool isOnline;
-
-  Location([
-    this.rsPeerId,
-    this.rsGpgId,
-    this.accountName,
-    this.locationName,
-    this.isOnline = false,
-  ]);
-}
