+++ lib/ui/change_identity_screen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:retroshare/model/identity.dart';
 import 'package:retroshare/provider/Idenity.dart';
 import 'package:retroshare/common/styles.dart';
 import 'package:retroshare/common/bottom_bar.dart';
 import 'package:retroshare/common/person_delegate.dart';
+import 'package:retroshare_api_wrapper/retroshare.dart';
 import 'package:shimmer/shimmer.dart';
 
 class ChangeIdentityScreen extends StatefulWidget {
