+++ lib/ui/create_room_screen.dart
@@ -3,14 +3,14 @@ import 'package:provider/provider.dart';
 import 'package:retroshare/common/input_chips/chips_input.dart';
 import 'package:retroshare/common/show_dialog.dart';
 import 'package:retroshare/common/styles.dart';
-import 'package:retroshare/model/location.dart';
 import 'package:retroshare/provider/friends_identity.dart';
 import 'package:retroshare/provider/Idenity.dart';
 import 'package:retroshare/provider/friend_location.dart';
 import 'package:retroshare/provider/subscribed.dart';
 import 'package:retroshare/common/person_delegate.dart';
-import 'package:retroshare/model/identity.dart';
-import 'package:retroshare/services/chat.dart';
+import 'package:retroshare/HelperFunction/chat.dart';
+
+import 'package:retroshare_api_wrapper/retroshare.dart';
 
 class CreateRoomScreen extends StatefulWidget {
   @override
