+++ lib/ui/discover_chats_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:retroshare/common/styles.dart';
 import 'package:retroshare/provider/subscribed.dart';
-import 'package:retroshare/services/chat.dart';
+import 'package:retroshare/HelperFunction/chat.dart';
 
 class DiscoverChatsScreen extends StatefulWidget {
   @override
@@ -59,14 +59,14 @@ class _DiscoverChatsScreenState extends State<DiscoverChatsScreen> {
                   ),
                   Expanded(
                     child: _chatsList.unSubscribedlist != null &&
-                            _chatsList.unSubscribedlist.length == 0
+                            _chatsList.unSubscribedlist.length > 0
                         ? ListView.builder(
                             padding: EdgeInsets.all(8),
-                            itemCount:1,
+                            itemCount:_chatsList.unSubscribedlist.length,
                             itemBuilder: (BuildContext context, int index) {
                               return GestureDetector(
                                 onTap: () {
-                                  /*_goToChat(_chatsList.unSubscribedlist[index]);*/
+                                  _goToChat(_chatsList.unSubscribedlist[index]);
                                 },
                                 child: Container(
                                   height: personDelegateHeight,
@@ -82,37 +82,37 @@ class _DiscoverChatsScreenState extends State<DiscoverChatsScreen> {
                                             crossAxisAlignment:
                                                 CrossAxisAlignment.start,
                                             children: <Widget>[
-                                              Text("djsdb",
-                                                /*_chatsList
+                                              Text(
+                                                _chatsList
                                                     .unSubscribedlist[index]
-                                                    .lobbyName,*/
+                                                    .lobbyName,
                                                 style: Theme.of(context)
                                                     .textTheme
                                                     .body2,
                                               ),
                                               Visibility(
-                                                visible: true,
-                                                /*_chatsList
+                                                visible: 
+                                                _chatsList
                                                     .unSubscribedlist[index]
                                                     .lobbyTopic
-                                                    .isNotEmpty*/
+                                                    .isNotEmpty,
                                                 child: Text(
-                                                  'Topic: ' ,
-                                                     /* _chatsList
+                                                  'Topic: ' +
+                                                     _chatsList
                                                           .unSubscribedlist[
                                                               index]
-                                                          .lobbyTopic,*/
+                                                          .lobbyTopic,
                                                   style: Theme.of(context)
                                                       .textTheme
                                                       .body1,
                                                 ),
                                               ),
                                               Text(
-                                                'Number of participants: ' /*+
+                                                'Number of participants: ' +
                                                     _chatsList
                                                         .unSubscribedlist[index]
                                                         .totalNumberOfPeers
-                                                        .toString()*/,
+                                                        .toString(),
                                                 style: Theme.of(context)
                                                     .textTheme
                                                     .body1,
@@ -125,8 +125,8 @@ class _DiscoverChatsScreenState extends State<DiscoverChatsScreen> {
                                         child: IconButton(
                                           icon: Icon(Icons.input),
                                           onPressed: () {
-                                           /* _goToChat(_chatsList
-                                                .unSubscribedlist[index]);*/
+                                            _goToChat(_chatsList
+                                                .unSubscribedlist[index]);
                                           },
                                         ),
                                       ),
