+++ lib/ui/launch_transition_screen.dart
@@ -2,7 +2,6 @@ import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'dart:io';
 
-import 'package:retroshare/services/account.dart';
 
 class LaunchTransitionScreen extends StatefulWidget {
   @override
