+++ lib/ui/room/message_delegate.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
-import 'package:retroshare/model/chat.dart';
+import 'package:retroshare_api_wrapper/retroshare.dart';
 
 class MessageDelegate extends StatelessWidget {
   const MessageDelegate({this.data, this.bubbleTitle});
