+++ lib/ui/room/messages_tab.dart
@@ -4,14 +4,14 @@ import 'package:emoji_picker/emoji_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:openapi/api.dart';
 import 'package:provider/provider.dart';
+import 'package:retroshare/HelperFunction/chat.dart';
 import 'package:retroshare/common/styles.dart';
 import 'package:retroshare/provider/room.dart';
 import 'package:retroshare/ui/room/message_delegate.dart';
 import 'package:retroshare/common/bottom_bar.dart';
-import 'package:retroshare/model/chat.dart';
-import 'package:retroshare/services/chat.dart';
+import 'package:retroshare_api_wrapper/retroshare.dart';
+
 
 class MessagesTab extends StatefulWidget {
   final Chat chat;
