+++ lib/ui/room/room_friends_tab.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:retroshare/model/identity.dart';
+import 'package:retroshare/HelperFunction/chat.dart';
 import 'package:retroshare/provider/friends_identity.dart';
 import 'package:retroshare/provider/room.dart';
 import 'package:retroshare/common/person_delegate.dart';
-import 'package:retroshare/model/chat.dart';
-import 'package:retroshare/services/chat.dart';
+import 'package:retroshare_api_wrapper/retroshare.dart';
+
 
 class RoomFriendsTab extends StatefulWidget {
   final Chat chat;
